Write checkpoint when end simulation time has been reached (#1218)
Description
This small change ensures that a checkpoint is written if the end simulation time has been reached regardless of the frequency for the check pointing specified in the parameter file.

Testing
All test involving a restart should pass without any issue. I have tested the feature with an example on my local machine and it works as expected.

Documentation
No change needed in the documentation.