Uniformize timer labels (#1220)
Description
The labels for the timer outputs where a bit all over the place. The majority of them follow the format "This is a sentence", but some previous output followed the syntax "this_is_a_sentence". I have made the labels uniform following the "This is a sentence" format. This is a very minor change towards a more unified Lethe.