Remove gear3 integrator (#1221)
Description
The Gear3 integrator was not used anywhere in the code in any example or in any tests. This integrator never gave good performance and was actually not tested for years. I believe it is now time to fully delete it. I want to work on adding new integrators this year and this will require some more unification. I want to start by cleaning things before adding new things.

Testing
There were actually no test for gear3...

Documentation
I adapted the documentation but the gear3 integrator was never mentioned in the documentation.