Rename all block Navier Stokes files using Fluid Dynamics prefix (#1226)
Description
This PR changes the name of the main class and the name of the files related to the lethe-fluid-block application. Is the third PR on a series of PRs that aim to rename all the Navier Stokes solvers of Lethe using the prefix FluidDynamics and removing GLS or GD.

Testing
All the existent tests pass without any change.

Former-commit-id: f2d74f13aae6fc7f1b9acf99c90d295c1bccf5bc