Rename all Nitsche files using Fluid Dynamics prefix (#1228)
Description
This PR changes the name of the main class and the name of the files related to the lethe-fluid-nitsche application. Is the fourth PR of a series of PRs that aim to rename all the Navier Stokes solvers of Lethe using the prefix FluidDynamics and removing GLS.

Testing
All the existent tests pass without any change.

Former-commit-id: de2a7b7e14652ed7770496315a386fa6ca13ee5a