Initialized previous_mesh_controller and previous number of elements to 0 in the constructor of the mesh controller object (#1230)
Description
Initialized previous_mesh_controller and previous number of elements to 0 in the constructor of the mesh controller object. previous_mesh_control_error which controls the intregral part of the error was not initialized. This affected geatly the PID controller convergence because really high values would sometimes be attributed at start to previous_mesh_control_error.

Solution
Initialized the value in the constructor.

Testing
old tests were kept