Rename all Sharp files using Fluid Dynamics prefix (#1231)
Description
This PR changes the name of the main class and the name of the files related to the lethe-fluid-sharp application. Is the fifth PR of a series of PRs that aim to rename all the Navier Stokes solvers of Lethe using the prefix FluidDynamics and removing GLS.

Testing
All the existent tests pass without any change.