Fix pressure interpolation at face bug and other bugs (#1234)
Description
The new deal.II changes (more asserts) identified a nasty bug we had. The wong fe_values was being used to gather the pressure at the faces. Instead of the fe_face_values it was the fe_values. Consequently, I am really not sure what was happening there. This is now fixed.

Furthermore, there were also instances where the quadrature of the fluid dynamics in the VOF solver was not initialized with the right number of points. This has been fixed also.

Testing
Tests that were breaking in debug should work just fine now in debug.