Ensure compilation works with dealii distributed vectors (#1239)
Description
Compilation with deal.II distributed vectors did not work anymore because of an addition in the CHN equation. This has been fixed. To ensure that this does not happen again I have added a CI step to test LDV compilation. This is a first step in removing Trilinos Vectors from Lethe.

Former-commit-id: 93ec65fb8b2e609d787ac0d175339ca959de0135