Print parameters (lethe-fluid-matrix-free) (#1255)
Print all relevant parameters at the end of the simulation. We can add a parameter to enable/disable this, but having the simulation output and the chosen parameters next to each other/in the same file is quite useful in my opinion. If you run too many parameter studies you loose the overview of the modified parameters.

Former-commit-id: 4f256b3d34569661896121f280b2a1230f6dce68