Specify GMRES parameters whent iterating intermediate GMG levels (#1264)
Description
In PR #1211 a new parameter was introduced that allow to iterate intermediate GMG levels. This allows to choose a level as coarse grid and used there a GMRES iterative solver preconditioned by a MG. So far, only one parameter was present: mg int level, however, the parameters for the number of iterations and tolerances were being hardcoded. This PR changes that by using already existing GMRES parameters..

Testing
I have tested this in the flow around a sphere case and seems to work just fine.

Documentation
The documentation was updated by adding a note that specifies the use of these parameters for this purpose as well.

Former-commit-id: bf39a0d6a199ddb258263132f9045a46a318d8f0