Add parameter to rescale mesh once it has been read from input files (#1276)
Description
Sometimes we read mesh files and rapidly realize that the simulation would be done more easily in a different set of unit (such that for example, the characteristic length of the mesh becomes 1 instead of 0.001). This is annoying since it requires regenerating the mesh, which is not always a trivial task. This PR adds a scaling parameter for every mesh read that allows a user to rapidly rescale a mesh by a constant factor > 0. This uses the deal.II GridTools.

Testing
To test the new feature I have added an application_test which is the traditional taylor-couette analytical solution, but where the mesh is actually rescaled once it has been established. This works well.

Documentation
Parameter has been added top the documentation.

Former-commit-id: bb3687095e31a2051b592db7dd9d8d5f2bd1ebaf