Function boundary conditions weakly imposed in the matrix-free solver (#1275)
Description
The matrix-free solver did not support weak function boundary conditions nor outlet. This is because we did not have the infrastructure to implement these features in the past. This PR adds the function weak capability to the matrix free solver. A PR will follow in this month to add the outlet boundary condition, making the matrix free solver feature complete for single phase newtonian flow.


Testing
I have added two application test to test the feature. The first one is an MMS with a single weak boundary condition imposed to zero. The second one is a taylor-couette which features mixed boundary conditions (regular no slip and function based)

Documentation
No additional documentation are required. However, I have also added assert so that the lethe-fluid-matrix-free solver throws when an invalid boundary condition is used.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Peter Munch <peterrmuench@gmail.com>
Former-commit-id: 7b9833b0878ae7141e07df4a77e213e7b47a0698