Fixed bug with weak boundary function evaluation (#1281)
Description
When multiple boundary conditions are presents, not all boundary have a function that is initialized. Currently, the matrix_free architecture still tried to evaluate functions for all boundaries at the face. This is not a good idea (and it is anyway completely useless). This PR ensures that functions are only evaluated if they are function_weak boundary condition (aka it the function really needs to be evaluated).