Add new scaling parameters for the Cahn-Hilliard model (#1274)
Description

With new models come new parameters that need to be dimensionalised adequately when changing the units in the "Dimensionality" section of the parameter file. With the Cahn-Hilliard model, it was the mobility coefficient, the interface thickness and the surface tension. Three new scalings were added at first, for those three parameters. Since surface tension is also used in VOF model, a fourth scaling parameter was added for the surface tension gradient for completion.

There was a mistake in the scaling of thermal expansion which was corrected.

Testing

No testing because it would imply coding an output for the physical properties which is unnecessary. It works fine.


Former-commit-id: 46511be4421d895a2ee5254fcf1e8558e9eb4de8