Fix bdf extrapolation bug (#1286)
Description
A bug in the VOF auxiliary physics was discovered when rerunning the gravity wave example. In the PR #994, BDF extrapolation of the velocity vector field in the VOF auxiliary physic was moved to the scratch data. However, the implementation was bypassed by an "if" condition.

Solution
This bug is fixed with this PR, and an application test was added to ensure that the feature remains intact with future implementations.

Testing
The new application test: applications_tests/lethe-fluid/vof-velocity-extrapolation.prm

Former-commit-id: 202fb53dbf68cc3379b7184cc0f44489d76eef50