Add outlet boundary condition to the matrix free operator (#1287)
Description
The lethe-fluid-matrix-free application did not support the outlet boundary condition. This is now integrated into the operators. The boundary condition was implemented using the functions created for the weak Dirichlet boundary conditions in a previous PR (#1275) . These functions were generalized and renamed so that they can be used for any boundary condition that requires the evaluation of face terms.

Testing
There was no test for the outlet BC. A new test has been added for both the matrix-free and the matrix-based application using the mesh cylinder_close_outlet.msh:

applications_tests/lethe-fluid-matrix-free/cylinder_outlet_bc.prm
applications_tests/lethe-fluid/cylinder_outlet_bc.prm

Documentation
A small note has been added to the documentation specifying which boundary conditions are currently not supported with the matrix-free application.