Correct update of solutions for multiphysics in matrix-free application (#1294)
Description
The matrix free application was updating the results outside of the iterate loop. This works only if the other physics are solved before the fluid dynamics, but we have other physics that are solved right after the fluid dynamics and in that case, they need the most recent fluid information within the iterate loop.

Solution
In order to fix this, we have to update the solutions right after solving the problem for the fluid dynamics, i.e., within the iterate loop right after solve_non_linear_system is called. For this, a function had to be created in the parent class NavierStokesBase.

Testing
No test available at the moment, the behaviour of the matrix-free application with the other physics is still a work in progress.