Correct Jacobian in tracer DCDD for transient simulations (#1297)
Description
The tracer matrix assembler was using the current tracer gradient in the Jacobian matrix but not in the residual for the calculation of the artificial shock capture viscosity. This is very bad because the Jacobian matrix becomes wrong and it is unclear if the solver will converge in this case. Furthermore, in the case of transient simulations, the solver was still trying to estimate the gradient of the vdcdd diffusivity, which is not meaningful because that gradient is zero since we are using the previous time step gradient.

Solution
This PR fixes this issue by using the correct Jacobian matrix implementation which uses the previous time step value of the gradient of the tracer for the DCDD calculation.