Installation instructions with faster cloning (#1301)
Description
The installation instructions using the git clone command without specifying that only the master branch needs to be pulled. This leads to people downloading the whole repository (800MB) vs the 120MB that would be sufficient for all their needs. Furthermore, the installation instructions mention that the p4est version of candi must be changed. This is not necessary anymore since candi is packaged by default with 2.3.6.

EDIT: The installation instructions also suggest using the package manager in ubuntu which is not a possibility right now since there are no 9.6 packages in ubuntu. I have fixed this part by commenting the code out and adding a warning.

I have changed the installation instructions to ensure that only the master branch is cloned with cloning. If someone needs other branches they can just check them out manually. Furthermore, I have removed the blocks about changing the version of p4est within candi.

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>