Fix simulation restart when using averaging (#1300)
Description
There was an issue where simulation restarts would crash on large parallel simulation when the time-averaging of the velocity was enabled. This issue was caused by the fact that the vectors that were checkpointed were the locally_relevant solutions but the restart was reading into the locally_owned solutions.

Solution
I have fixed this by reading the restart into the locally_relevant solutions. However, I had to add a function to manually sanitize the time_averaging operator because the locally_owned vectors were now empty.

Testing
This should not affect any of the current test results. I do not know if it is necessary to add a new tests since restart tests are cumbersome. 

Documentation
No changes to documentation.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>