Fix restart simulation time (#1302)
Description
Simulation restarted when time dependent boundary conditions were put to "on" had a bad initial solution because it would use the local evaluation point to reset the value. This is very bad, because it would essentially reset the evaluation point to 0 making convergence quasi impossible. This bug had been fixed in the matrix based solver, but not the matrix free.

Solution
I applied the same correction as I applied to the matrix based solver. Only this time I found it much quicker.

Testing
I added a new test. The new test is a bit weird, but quite simple. It's a turbulent taylor-couette simulation that has run for a lot of time steps (10.5 seconds with time step of 0.01). Then we restart the simulation at 10.5 and finish it at 10.6. Because the time dependent flag is set to true, in the previous version the initial guess would be very far away from the solution at 10.5.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>