Relicense to Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
And add SPDX-FileCopyrightText and SPDX-License-Identifier tags to most
files.

The new license is the same as the one all new deal.II code is licensed
under [1], and since Lethe is heavily based on deal.II, this simplifies
the exchange of code between the projects.

We have permission to relicense in this way all significant
contributions to Lethe [2].

[1]: https://github.com/dealii/dealii/blob/dealii-9.6/LICENSE.md
[2]: https://github.com/chaos-polymtl/lethe/issues/1145