Fix nomenclature of lagrangian post-processing (#1305)
Description
The lagrangian postprocessing parameter was our only parameter with a capital letter. This was a confusing decision because we use all lower case parameters. This pull request fixes this by making it a lower case "l" and adapting the documentation to follow this nomenclature.

Furthermore, we had an example ( rotating-drum-with-postprocessing) that was nothing more than the rotating-drum example, but with an additional parameter set to on. The example did not show any results or analysis. Consequently, I have decided to delete it in order not to confuse the readers. For the moment I have added the lagrangian post-processing to the regular rotating drum example. Once the example will be improve, we can actually use this to analyse the results.

This addresses #1176

Co-authored-by: OGaboriault <oliviergabo@gmail.com>