Dem fixing resize containers (#1316)
Description
Some std::vector<> in the DEM were not resize properly. The resize was being done before the exchange of ghost particles between processors, thus the max particle index was be too small. This was causing unexpected segmentation faults when cells had a lot of neighbouring cells. This would happen especially when dealing with periodic boundary condition.

Solution
The exchange of ghost particles must be done before the resize. This is a permanent fix.

Testing
A new test was added using the periodic hopper example. This test is currently crashing on master. It uses 2 processor, otherwise there is not ghost particles.

Documentation
N/A