Bug when using multiple manifolds (#1312)
Description
A bug was preventing the use of multiple manifolds. The variable describing the manifold was set as a static const but needed to be modified for every manifold.

The maximum number of manifolds was also set to 14. To make it more flexible, the method used for the boundary conditions was modified to take the manifolds into account.

Solution
removed the static const.