Fix TGV restart test for matrix-free (#1321)
Description
The TGV restart test used in the matrix-free application was using matrix-based restart files (generated with an L2 projection initial condition, which is not supported by matrix-free) and a GMRES coarse-grid solver.

Solution
New restart files were generated and the generator prm file was added. In addition, the coarse-grid solver used is now a direct solver.

Testing
The test passes in both debug and release mode.