Add Discontinuous Galerkin capabilities to the Tracer auxiliary physics (#1320)
Description
Adds the capability to the Tracer to use a Discontinuous Galerkin formulation instead of a Continuous Galerkin formulation. When DG is used, an entirely different paradigm has to be used to calculate the integral at the internal faces and at the boundary faces. This PR deploys this architecture as well as numerous application_test to test the DG implementation.

Concretely, here is a summary of what is done:

A new assembly routine for the RHS and the Matrix have been added to the tracer. This new assembly routine uses the MeshWorker paradigm to loop over inner faces, boundary faces and cells.
The tracer scratch data has been modified to accomodate the necessary DG features. This includes calculation of the velocity at the internal/boundary faces, calculation of the tracer terms at the faces and calculation of the physical properties at the faces.
An outlet boundary condition has been added for the tracer physics.
Assemblers for the DG equations have been added.
Testing
9 additional application_tests were introduce to fully test the DG implementation.

Documentation
Few modifications are required in the .RST files, they have been added.

Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>
Co-authored-by: mivaia <59896845+mivaia@users.noreply.github.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>