work in progress part 1
I have refactored the boundary conditions to be map inside of the
boundary_conditions.h file. This is a first step, but it means I need to
refactor everything all at once. It is going to be more work than
expected, but I believe it will be very positive in the end