Double restart  (#1327)
Description
Sometimes, the file system on a cluster can be slow. In the last few month, we had many simulations where the allocated time expired while the checkpoint files were being written. This made the restart files corrupted and those simulations had to be restart from the beginning, which is annoying.

This feature make make code alternates between two different version of the restart files (_0. and _1. ) this way, there is always at least one version of restart file that is not corrupted.

Currently, it only works for DEM simulation, but will be implemented for every solver. Also, we could add a new parameter to specify which restart version to use.

Testing
I don't think it's necessary to add new tests, since the existing tests using restart had to be modified for this feature.