Use map to store boundary conditions instead of vectors (#1330)
Description
The way we stored boundary condition was always poorly designed. We used a vector with ids that would correspond to the deal.II IDs, but we also had the lethe-ids which were contiguous. This was very confusing (which id is supposed to be what?) and it also meant we always had to search through the vector of ids.

I refactored this entirely to instead use a map where the key is the boundary id and the values is the type (or the function, or etc). This way, we never have to look up what is "our id", "our id" is always the ID we put in the mesh generator for the boundary condition.

This required some extensive refactoring, but in the end i find that the final design is much much cleaner.

Testing
This does not include new test. I had to remake ONE restart file for ONE test. That is the fluidized bed case. The issue is that the order in which we apply the boundary conditions changed in this case and so that actually change the reslts because the corner nodes for the velocity are a bit difference. I used this opportunity to make the test a bit better and more reproducible.

Documentation
Nothing to document

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>