Improve cylinder example
I have fixed the cylinder example to talk about the outlets. I have also
removed statements regarding manifolds that were not adequate anymore.
Finally, I realized that the geo file used to generate the mesh for this
case made no sense since it generated a mesh that was far too fine for
the simulation to be carried out. I have thus made the geo file
consistent with the content of the example