CHNS example : capillary rise and Jurin's law (#1328)
Description
This is the first example of the Cahn-Hilliard-Navier-Stokes solver. It displays the use of angle of contact boundary conditions to solve a capillary rise problem in 2D. The numerical height difference is compared to that predicted by the Jurin's law and show excellent agreement . The pressure fields are qualitatively commented and show that the solver adequately describe pressure jumps due to an interface.

A post processing python script is included with the example to reproduce the figures if needed. They were added in a previous PR (#1307) and already reviewed.

EDIT
There was a modification made to the implementation of the boundary conditions to make the example work with the new boundary conditions.
An application test (cahn_hilliard_navier_stokes_jurins_law_2d) was added to ensure of the future stability of the solver

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>