Improve cavity example (#1340)
Description
The lid-driven cavity example was just fine, but there was a small issue with the Reynolds 7500 case. The tolerance was so low that it took forever to converge. So I increased it, the results don't change. However, this made me realize that the python script was hardcoded to use a given file number. I fixed this by using a regular expression to find the last file no matter what.

Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>