Add new sharp-edge example: Drag around a cylinder (#1299)
Description
This example showcases the use of sharp immersed boundary conditions for the evaluation of the flow around a non-spherical particle. We show the validity of the method at evaluating the drag on the particle with values from litterature. For now the example targets a Reynolds range of 0.1 to 1with low aspect ratio particles (2) in cross flow.


Co-authored-by: JustinLamouche <justin.lamouche@polymtl.ca>
Co-authored-by: JustinLamouche <135653039+Julamipolymtl@users.noreply.github.com>
Co-authored-by: Pierre Laurentin <81017102+PierreLaurentinCS@users.noreply.github.com>