Improve cylinder example (#1337)
Description
I have fixed the cylinder example to talk about the outlets. I have also removed statements regarding manifolds that were not adequate anymore. Finally, I realized that the geo file used to generate the mesh for this case made no sense since it generated a mesh that was far too fine for the simulation to be carried out. I have thus made the geo file consistent with the content of the example

Documentation
The documentation has been updated accordingly.

Co-authored-by: Pierre Laurentin <81017102+PierreLaurentinCS@users.noreply.github.com>