Update flow around sphere example (#1346)
Description
The example needed to be verified. The new boundary condition was added to the schematic of the case, minor typos were corrected, some sentences were rephrased, and the initial refinement parameter was missing from one of the parameter files. After the changes all the results presented in the doc page correspond to the ones obtained with the parameter files given in the examples folder.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>