Packing in circle example (#1350)
Description
The packing in circle example was a bit of a mess. It contained parameters which were deprecated a long time ago. I have cleaned it. I also realized the time step that was being used was far too small and increased the time-step. The tutorial now runs in like 7s.



Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>