Update transient flow around cylinder example (#1348)
Description
The transient flow around a cylinder example works very well. However, some small things and a typo needed to be fixed.

A) Keeping the .msh file inside the folder is not a good idea I think. We have moved away from this. I have removed the file but kept the .geo file

B) I have added instructions on how to use the .geo file to make the mesh

C) The AMG preconditioner does not need additional parameters, the default values are fine.

D) There was a weird typo on the CL values, there was a zero too much there. I have fixed that.

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>