CHNS example : bubble detachment in a shear flow (#1334)
Description
This is the second example of the Cahn-Hilliard-Navier-Stokes solver. The solver is used to simulate the detachment of bubbles in a shear flow in three dimensions. The quantities of interest (detachment time and volume) are compared to values from the literature. An excellent agreement is observed, which confirms that the CHNS solver is able to describe complex interface physics, involving break-up and jet formation.

All the necessary files to run the parametric sweep are included (only on the shear rate) and python scripts are provided to reproduce the figures if needed (see the previous PR : #1307).