Update TGV example in documentation (#1363)
Description
This PR updates the parameters of both the matrix-based and matrix-free TGV simulations according to the findings in the matrix-free article. To summarize:

BDF3 scheme was changed to BDF2.
Linear solver parameters were already tuned for the article, so we updated tolerances, number of iterations and krylov vectors accordingly.
The example was written with constant time step, which works well so it won't be changed. However, trying constant CFL was added as a suggestion and the script to calculate the kinetic energy rate for that case was already available in the example folder.
I do not think it is necessary to re run all the cases, so the part that talks about approximate times was left as it is.