Fix bugs: changing time step after restart and tracer tanh model (#1343)
Description
There were two issues that appeared when using the tracer physics with an immersed solid:

Evaluation of the diffusivity model immersed solid tanh was too costly
Changing the time step after restarting was impossible
Solution
The diffusivity model now uses the optimizations from the Shape class, that allow to evaluate the distance using a cell guess; this makes RBF and composite distance evaluations magnitudes faster.
Time step change after a restart is now allowed when adapt==false. This behavior is not allowed when adapt==true, as then we desire a continuity in time step evolution. When using fixed time step though, users who change the value after a restart apparently aim for this behavior.


Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>