update ball-packing and rectangular-hopper examples and doc (#1374)
Description
Update of the .prm and .rst files of the ball-packing and rectangular-hopper examples for the Lethe 1.0 release. The rectangular hopper gave similar curves visually, but the mass flow rate is different than the one stated in the documentation (85.17 vs 84.94 & 90.12 vs 85.09 g/s). Nonetheless, I feel that those results are close enough that it is not necessary to change the example documentation for it.