Cylindrical spouted-bed: Improve documentation of the example and fix typos (#1373)
Description
This PR improves the documentation for the cylindrical spouted bed example. The example works very well, but the timing reported in the documentation were completly off (like so off it makes no sense. For example it was reported that the example would take 8 days to run, it took 5 hours on my computer while I was doing other stuff and other simulations at the same time). I have simplified the wording of the example and improved the discussion. The animations and etc. I have not touched, they are already great as they are. The example works fine as is though.

Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>