NACA0012 example review (#1375)
Tested:

The .geo file to generate .msh files -> it works perfectly. Different NACA angles can be generated by changing a single parameter.
The .prm file. The simulation is REALLY expensive (10% of a single simulation in 15h on 15 cores). Results match perfectly after 10% of the simulation.
The post-processing file. It has been modified to add the reference correlation in the fast Fourier transform analysis.