Cylindrical packed bed (#1369)
Description
The steady-state flow withing a cylindrical packed bed example was working fine, but the documentation was outdated. Furthermore, the figures we atrociously bad and were not representative enough. I have updated the example results and I have added new figures which illustrate the behavior of the VANS equation in a better fashion. I have also added a simple python post-processing script to do these calculations.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>