Test without void fraction bounding (#1386)
Description
As part of the current void fraction refactoring I wish to aim at deprecating the void fraction bounding after the refactoring is done. This requires modifying the test that currently use void fraction bounds. This PR updates the tests that use lethe-fluid-vans to not bound the void fraction. This is all it does, update the tests.

Testing
Test results are updated.

Documentation
Nothing to be done here