Update flow past a backward facing step example (#1389)
Description
The post-processing scripts were not working for this example and the example was long and hard to follow. This PR refactors the example by limiting it to two Reynolds numbers: 100 and 1000. To summarize:

All the parameters are revisited to ensure correct convergence for both cases.
The folders and parameter files are renamed so that it is clear for which Reynolds they are.
The results are limited to: visualizing the velocity profiles (with streamlines to identify eddy after step and zoom in plots, all in one figure), reporting the reattachment point for the last step and its relative error in comparison to reference data, and comparing the obtained velocity distribution at the outlet to the analytical solution.
The post-processing scripts are fixed so that they work for both Reynolds number by just changing the appropriate flag when running the script. Moreover, the number of vtu files is no longer hardcoded and can be found automatically by the script.

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>