Rotating drum example (#1391)
Description
I ran the rotating drum example. The only thing I have changed in the prm file is the post-processing Lagrangian. Since it is set at false in the loading simulation, the code is crashing in the second since it is looking for the pvd associated with the post-processing lagrangian.